style: format and reorganize files
