+++ src/main/kotlin/ecommerce/products/exception/ProductValidationException.kt
@@ -1,3 +1,3 @@
-package ecommerce.exception
+package ecommerce.products.exception
 
 class ProductValidationException(message: String) : RuntimeException(message)
